+++ src/section/templates/parsers/index.js
@@ -1,9 +1,14 @@
 // const parseCitation = require('./citation');
 const keyValue = require('./key-value');
 
-const grabInside = function(tmpl) {
+const trim = function(tmpl) {
   tmpl = tmpl.replace(/^\{\{/, '');
   tmpl = tmpl.replace(/\}\}$/, '');
+  return tmpl;
+};
+
+const grabInside = function(tmpl) {
+  tmpl = trim(tmpl);
   let parts = tmpl.split('|');
   if (typeof parts[1] !== 'string') {
     return null;
@@ -27,6 +32,22 @@ const parsers = {
     return {
       template: 'citation',
       data: data
+    };
+  },
+  //this one sucks - https://en.wikipedia.org/wiki/Template:GNIS
+  'cite gnis': (tmpl, options) => {
+    if (options.citations === false) {
+      return null;
+    }
+    tmpl = trim(tmpl);
+    let arr = tmpl.split('|');
+    return {
+      template: 'citation',
+      data: {
+        id: arr[1],
+        name: arr[2],
+        type: arr[3],
+      }
     };
   }
 };