+++ changelog.md
@@ -53,3 +53,4 @@
 * `.templates()` are now an ordered array, instead of an object, and include infoboxes and citations
 * add (early) support for 'generic' key-value template parsing
 * normalize/lowercase template/infobox properties - add loose `.get('key')` method to Infobox class
+* mess-around with citation-template formatting